all: unsigned cursor tests, XOR digit mask, word-folded integer readers
A hardware-counter pass on a Neoverse N2 (Azure Cobalt-class VM, perf stat
per-op differencing over the 13 PMU events a Hyper-V guest exposes). The
corpus is issue-bound here: 24 of 30 cases run at IPC 3.6-4.6 with under
0.5% branch misses, so executed instructions decided every keep and revert.
The Meteor Lake commit before this one was validated on N2 first (canada
-34%, numbers -28% instructions per decode, geomean -7.2%).

- Every cursor bound in the generated decoders and the runtime is tested
  unsigned (uint(i) < uint(len(data))). The compiler cannot prove an int
  cursor non-negative once it has passed through reader results, so each
  data[i] behind a signed test kept a second, never-taken branch to the
  bounds panic - seven per object member, 2-4% of instructions on the
  object and batch cases, one dispatch slot each on an issue-bound core.
- The SWAR digit test is XOR-based with bitmask-immediate constants only:
  d = w ^ 0x30.., ((d + 0x06..) | d) & 0xf0... The subtract-and-bias form
  materialised a four-instruction splat per scanFloat call, and the
  compiler distributed the fold's multiply over the subtraction, adding a
  two-instruction constant on the fraction path. 1-4% fewer instructions on
  the number cases; the exhaustive tryParse4Digits oracle and the fast/slow
  float differential pass unchanged.
- ReadInt64OrNull/ReadUint64OrNull fold their digit run a word at a time
  (digitRun): golang_source -5.5%, citm -3.5% instructions. Deliberately not
  applied to the batch array loops, where it measured mesh +1.1%: a byte
  loop's loads issue under a predicted branch, a word count puts the next
  element's address on a data chain. TestDigitRunWordBoundary pins the run
  that ends exactly on a word boundary, which the first version got wrong.

Full 30-case corpus vs the previous commit, interleaved ABBA, n=8, both
sides linked -funcalign=64: geomean -2.4%; golang_source -6.7%, numbers
-6.7%, mesh -5.5%, marine_ik -5.0%, float-array -4.3%, canada -3.8%, citm
-3.5%, twitter -3.2%, cloudflare -2.4%; six cases flat; string_unicode
+0.3% (p=0.037) the one detectable regression. A Go SWAR probe in front of
the SVE2 string scanner was built three ways and rejected (it bets on the
document's key lengths: cloudflare +15%); CLAUDE.md records the method, the
per-case counter table, the entries and the rejection.

+++ main.go
@@ -763,13 +763,24 @@ func (g *gen) skipWS(dst, src string) string {
 	// the SWAR unstable.SkipWSRun, which eats it eight bytes at a time. Doing this
 	// at the call site rather than in unstable.SkipWS sidesteps the inliner budget
 	// that would otherwise force every skip through a function call.
+	//
+	// Every bound here and in the loop templates below is tested UNSIGNED,
+	// uint(i) < uint(len(data)). The cursor is an int that has passed through
+	// reader results, so the compiler cannot prove it non-negative, and after a
+	// signed i < len(data) every data[i] kept its own bounds check: on arm64 a
+	// second, never-taken branch to panicBounds per probe (the compare's flags
+	// are reused; on amd64 it is a compare and a jump). The unsigned form proves
+	// 0 <= i < len(data) in one test and is identical for every i >= 0, which is
+	// every i the generated code ever holds. Measured on a Neoverse N2 (2026-09):
+	// 2-3% fewer instructions per decode and 8-10% fewer branches on the object
+	// cases; time citm_catalog -2.7%, twitter_status -1.7%, cloudflare -1.3%.
 	set := ""
 	if dst != src {
 		set = dst + " = " + src + "\n\t"
 	}
-	return fmt.Sprintf(`%[1]sif %[2]s < len(data) && data[%[2]s] <= ' ' {
+	return fmt.Sprintf(`%[1]sif uint(%[2]s) < uint(len(data)) && data[%[2]s] <= ' ' {
 		%[2]s++
-		if %[2]s < len(data) && data[%[2]s] <= ' ' {
+		if uint(%[2]s) < uint(len(data)) && data[%[2]s] <= ' ' {
 			%[2]s = unstable.SkipWSRun(data, %[2]s+1)
 		}
 	}`, set, dst)
@@ -792,11 +803,11 @@ func (g *gen) skipWS(dst, src string) string {
 func (g *gen) readKey() string {
 	return `var key string
 	var ni int
-	if i >= len(data) || data[i] != '"' {
+	if uint(i) >= uint(len(data)) || data[i] != '"' {
 		return i, unstable.ErrInvalidJSON
 	}
 	ks := i + 1
-	if ke := unstable.IndexCloseOrEscapeAt(data, ks); ke < len(data) && data[ke] == '"' {
+	if ke := unstable.IndexCloseOrEscapeAt(data, ks); uint(ke) < uint(len(data)) && data[ke] == '"' {
 		key, ni = unstable.UnsafeStr(data[ks:ke]), ke+1
 	} else {
 		var err error
@@ -910,7 +921,7 @@ func (g *gen) genUnmarshal(name string) string {
 	// the README's directive documentation instead.
 	return fmt.Sprintf(`func (v *%[1]s) UnmarshalJSON(data []byte) error {
 	%[3]si := unstable.SkipWS(data, 0)
-	if i >= len(data) {
+	if uint(i) >= uint(len(data)) {
 		return unstable.ErrTruncated
 	}
 	if data[i] == 'n' {
@@ -1323,7 +1334,7 @@ func (g *gen) genStructBody(fn, paramType string, st *ast.StructType) {
 	// after each member only) measured cloudflare +11% — the wide decoder is
 	// that layout-sensitive — so don't restructure, flag.
 	body := fmt.Sprintf(`func %[1]s(v %[2]s, data []byte, i int%[9]s) (int, error) {
-	%[10]sif i >= len(data) {
+	%[10]sif uint(i) >= uint(len(data)) {
 		return i, unstable.ErrTruncated
 	}
 	if data[i] == 'n' {
@@ -1335,7 +1346,7 @@ func (g *gen) genStructBody(fn, paramType string, st *ast.StructType) {
 	i++
 	for first := true; ; first = false {
 		%[4]s
-		if i >= len(data) {
+		if uint(i) >= uint(len(data)) {
 			return i, unstable.ErrTruncated
 		}
 		if data[i] == '}' {
@@ -1346,16 +1357,16 @@ func (g *gen) genStructBody(fn, paramType string, st *ast.StructType) {
 		}
 		%[8]s
 		%[5]s
-		if i >= len(data) || data[i] != ':' {
+		if uint(i) >= uint(len(data)) || data[i] != ':' {
 			return i, unstable.ErrExpectColon
 		}
 		%[6]s
-		if i >= len(data) {
+		if uint(i) >= uint(len(data)) {
 			return i, unstable.ErrTruncated
 		}
 		%[3]s
 		%[7]s
-		if i >= len(data) {
+		if uint(i) >= uint(len(data)) {
 			return i, unstable.ErrTruncated
 		}
 		if data[i] == '}' {
@@ -1896,7 +1907,7 @@ if berr != nil {
 if len(body) > 0 {
 	if _, ierr := func(data []byte, i int) (int, error) {
 		i = unstable.SkipWS(data, i)
-		if i >= len(data) {
+		if uint(i) >= uint(len(data)) {
 			return i, nil
 		}
 %s
@@ -1951,7 +1962,7 @@ func (g *gen) arrayDecoder(t *ast.ArrayType, hint string, nocopy, lax bool) stri
 	// Trailing commas are rejected by the first-iteration flag, as in
 	// genStructBody.
 	body := fmt.Sprintf(`func %[1]s(out *%[2]s, data []byte, i int%[5]s) (int, error) {
-	if i >= len(data) {
+	if uint(i) >= uint(len(data)) {
 		return i, unstable.ErrTruncated
 	}
 	if data[i] == 'n' {
@@ -1965,7 +1976,7 @@ func (g *gen) arrayDecoder(t *ast.ArrayType, hint string, nocopy, lax bool) stri
 	idx := 0
 	for first := true; ; first = false {
 		%[3]s
-		if i >= len(data) {
+		if uint(i) >= uint(len(data)) {
 			return i, unstable.ErrTruncated
 		}
 		if data[i] == ']' {
@@ -1985,7 +1996,7 @@ func (g *gen) arrayDecoder(t *ast.ArrayType, hint string, nocopy, lax bool) stri
 		}
 		idx++
 		%[3]s
-		if i >= len(data) {
+		if uint(i) >= uint(len(data)) {
 			return i, unstable.ErrTruncated
 		}
 		if data[i] == ']' {
@@ -2168,7 +2179,7 @@ func (g *gen) sliceDecoder(elt ast.Expr, hint string, nocopy, lax, root bool) st
 	// (which would cost a heap allocation per element for slices of
 	// structs/pointers).
 	body := fmt.Sprintf(`func %[1]s(out *[]%[2]s, data []byte, i int%[7]s) (int, error) {
-	if i >= len(data) {
+	if uint(i) >= uint(len(data)) {
 		return i, unstable.ErrTruncated
 	}
 	if data[i] == 'n' {
@@ -2188,7 +2199,7 @@ func (g *gen) sliceDecoder(elt ast.Expr, hint string, nocopy, lax, root bool) st
 %[4]s	i++
 	for first := true; ; first = false {
 		%[5]s
-		if i >= len(data) {
+		if uint(i) >= uint(len(data)) {
 			return i, unstable.ErrTruncated
 		}
 		if data[i] == ']' {
@@ -2200,7 +2211,7 @@ func (g *gen) sliceDecoder(elt ast.Expr, hint string, nocopy, lax, root bool) st
 		%[6]s
 		%[3]s
 		%[5]s
-		if i >= len(data) {
+		if uint(i) >= uint(len(data)) {
 			return i, unstable.ErrTruncated
 		}
 		if data[i] == ']' {
@@ -2359,7 +2370,7 @@ func (g *gen) mapDecoder(keyExpr, valExpr ast.Expr, hint string, nocopy, lax boo
 	// Trailing commas are rejected by the first-iteration flag, as in
 	// genStructBody.
 	body := fmt.Sprintf(`func %[1]s(out *map[string]%[2]s, data []byte, i int%[9]s) (int, error) {
-	if i >= len(data) {
+	if uint(i) >= uint(len(data)) {
 		return i, unstable.ErrTruncated
 	}
 	if data[i] == 'n' {
@@ -2380,7 +2391,7 @@ func (g *gen) mapDecoder(keyExpr, valExpr ast.Expr, hint string, nocopy, lax boo
 	}
 	for first := true; ; first = false {
 		%[4]s
-		if i >= len(data) {
+		if uint(i) >= uint(len(data)) {
 			return i, unstable.ErrTruncated
 		}
 		if data[i] == '}' {
@@ -2392,18 +2403,18 @@ func (g *gen) mapDecoder(keyExpr, valExpr ast.Expr, hint string, nocopy, lax boo
 		}
 		%[7]s
 		%[5]s
-		if i >= len(data) || data[i] != ':' {
+		if uint(i) >= uint(len(data)) || data[i] != ':' {
 			return i, unstable.ErrExpectColon
 		}
 		%[6]s
-		if i >= len(data) {
+		if uint(i) >= uint(len(data)) {
 			return i, unstable.ErrTruncated
 		}
 		var val %[2]s
 		%[3]s
 		%[8]s
 		%[4]s
-		if i >= len(data) {
+		if uint(i) >= uint(len(data)) {
 			return i, unstable.ErrTruncated
 		}
 		if data[i] == '}' {
